+Move array_global_min_max to MOM_spatial_means
  Moved the publicly visible routine array_global_min_max to MOM_spatial_means
from MOM_generic_tracer, along with the private supporting function ijk_loc,
without any changes to the code itself.  This was done because this routine now
gives useful debugging information that can be useful outside of the generic
tracer module.  All answers are bitwise identical, but the module use statements
for array_global_min_max will need to be updated.