CI: Fortran 2018 testing
This patch enables Fortran 2018 standard compliance testing.  We could
do 2023, but our current CI of choice doesn't yet have a compiler which
can do this.

The actual content of this PR is a decoupling of FCFLAGS_DEBUG and
FCFLAGS_FMS.  There is now a default FCFLAGS macro which is used by the
other two macros.  One can now optionally configure FCFLAGS_DEBUG
without worrying about the impact on FCFLAGS_FMS.

The motivation here is that we don't want to test for F2018/2023
compliance in FMS.