Move and convert MOM6_generic_tracer to a stub
Move MOM_generic_tracer out of MOM6 and into the externals as a stub.
With this change, users of the generic tracer modules will be requied
to add an active version of MOM_generic_tracer to their equivalent of
the GFDL_ocean_BGC folder.

This should not impact users running without generic tracers.